optik-platform/backend/requirements.txt
# ============================================
# OPTIK PLATFORM - Backend Dependencies
# ============================================

# --- Core Framework ---
fastapi==0.111.0
uvicorn==0.30.1
pydantic>=2.7.2

# --- Database ---
# MongoDB (Primary)
motor>=3.3.2  # Async MongoDB driver
pymongo>=4.6.0  # MongoDB Python driver

# PostgreSQL (Alternative)
databases[asyncpg]>=0.9.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# --- Scraping & Parsing ---
beautifulsoup4>=4.12.3
httpx>=0.27.0
aiohttp>=3.9.5

# --- AI / LLM Agents ---
langchain>=0.2.1
langchain-anthropic>=0.1.15
anthropic>=0.25.0

# --- Blockchain (Solana) ---
solana>=0.33.0
solders>=0.20.0
anchorpy>=0.20.1
base58>=2.1.1

# --- Payments ---
stripe>=9.0.0

# --- Storage (IPFS) ---
# pinata-py==0.1.3  # Optional - uses direct API calls instead

# --- Security ---
boto3>=1.34.0  # AWS Secrets Manager
cryptography>=42.0.0  # Encryption utilities

# --- Utilities ---
python-dotenv>=1.0.1
reportlab>=4.2.0
aiosqlite>=0.17.0

PyJWT>=2.8.0
redis>=5.0.1
email-validator>=2.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-multipart>=0.0.9
